+++ app/controllers/events_controller.rb
@@ -1,6 +1,5 @@
 class EventsController < ApplicationController
   skip_before_action :require_login, only: [:index, :index_past, :show, :create, :new, :preview]
-  before_action :get_event, only: [:admin_show, :approve, :edit, :update, :destroy]
 
   def index
     @events = Event.approved.upcoming
@@ -9,21 +8,6 @@ class EventsController < ApplicationController
 
   def index_past
     @events = Event.approved.past
-  end
-
-  def admin_index
-    @categorized_events = {
-      "Unapproved Events" => Event.unapproved,
-      "Approved Events" => Event.approved.upcoming,
-      "Past Events"=> Event.approved.past
-    }
-  end
-
-  def admin_show
-    respond_to do |format|
-      format.html
-      format.csv { send_data @event.to_csv, filename: "#{@event.name} #{DateTime.now.strftime("%F")}.csv" }
-    end
   end
 
   def show
@@ -55,31 +39,6 @@ class EventsController < ApplicationController
     end
   end
 
-  def edit
-  end
-
-  def update
-    if @event.update(event_params)
-      redirect_to admin_path, notice: "You have successfully updated #{@event.name}."
-    else
-      render :edit
-    end
-  end
-
-  def approve
-    @event.toggle(:approved)
-    @event.save
-    if @event.approved?
-      OrganizerMailer.approved_event(@event).deliver
-    end
-    redirect_to admin_url
-  end
-
-  def destroy
-    @event.destroy
-    redirect_to admin_url
-  end
-
   private
     def event_params
       params.require(:event).permit(
@@ -87,9 +46,5 @@ class EventsController < ApplicationController
         :description, :name, :start_date, :end_date, :approved, :ticket_funded,
         :accommodation_funded, :travel_funded, :deadline, :number_of_tickets,
         :website, :code_of_conduct, :city, :country, :applicant_directions)
-    end
-
-    def get_event
-      @event = Event.find(params[:id])
     end
 end
