+++ app/helpers/admin_events_helper.rb
@@ -1,0 +1,2 @@
+module AdminEventsHelper
+end
