+++ test/controllers/admin_events_controller_test.rb
@@ -1,0 +1,12 @@
+require 'test_helper'
+
+class AdminEventsControllerTest < ActionController::TestCase
+  # test 'does not show admin stuff to logged-in non-admins' do
+  #   user = User.create
+  #   sign_in_as(user)
+
+  #   get :index
+
+  #   assert_redirected_to sign_in_path
+  # end
+end
